Create Meson build files
Tolk can now be built with Meson. The Meson build system builds
Tolk.dll, TolkDotNet.dll, and Tolk.jar, and is instructed on how to
distribute tolk.py, and build and distribute README.html.

Meson feature options are used to determine which bindings are built and
distributed. By default, they're set to auto, so bindings will
automatically be built if the corrisponding language compilers /
runtimes are found. There is also a feature option for docs, which will
build the documentation if Pandoc is found when set to auto. There is
also a boolean feature that determines whether JNI support is enabled.
If building Java bindings, this will be forced on.

This makes it significantly easier to use Tolk from Meson projects as a
subproject. It also means Tolk benefits from Meson's existing features
and compiler / language support and integration, so theoretically one
could more easily build Tolk with MinGW.